fix 4 BusStop läuft auf allen Schichten (Repository fehlt noch)

+++ be219thenextchapter-master/src/main/java/BE219Thenextchapter/Service/BusStopServiceImpl.java
@@ -8,6 +8,7 @@ import BE219Thenextchapter.repository.BusStopRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class BusStopServiceImpl implements BusStopService{
@@ -34,6 +35,9 @@ public class BusStopServiceImpl implements BusStopService{
     //Read by Id
     @Override
     public BusStopDTO getBusStopById(long id) {
+        if (!busStopRepository.findById(id).isPresent()) {
+            throw new NoSuchElementException();
+        }
         BusStop busStop = this.busStopRepository.findById(id).get();
         BusStopDTO busStopDTO = this.busStopMapper.busStopToBusStopDTO(busStop);
         return busStopDTO;
@@ -50,8 +54,12 @@ public class BusStopServiceImpl implements BusStopService{
 
     //Update
     @Override
-    public BusStopDTO updateBusStop(BusStopDTO busStopDTO) {
+    public BusStopDTO updateBusStop(long busStopId, BusStopDTO busStopDTO) {
+        if (!busStopRepository.findById(busStopId).isPresent()) {
+            throw new NoSuchElementException();
+        }
         BusStop busStop = this.busStopMapper.busStopDTOToBusStop(busStopDTO);
+        busStop.setId(busStopId);
         this.busStopRepository.save(busStop);
         return this.busStopMapper.busStopToBusStopDTO(busStop); //Umwandlung und Rückgabe des Objekts in DTO über die Mapper-Methode
         }
@@ -60,6 +68,9 @@ public class BusStopServiceImpl implements BusStopService{
     //Delete
     @Override
     public void removeBusStopById(long id) {
+        if (!busStopRepository.findById(id).isPresent()) {
+            throw new NoSuchElementException();
+        }
         this.busStopRepository.deleteById(id);
     }
 }
